acert: Add role attribute display to test
Explicitly print the group and role attributes found in the test
attribute certificate.  This is used to verify that the template
for the attribute structure is working as expeceted and that
the attribute can successfully be set from the configuration file.

This is a temporary workaround for the fact that there is no generic
attribute parsing like we have for extensions, but at least gives
some test coverage for the new attributes.